Optimize SimpleSubSetIterator.
This commit provides significant performance boost for subset generation.
Firstly, when assembling a next subset, namely instead of
looping over original input and checking a subset state in the bit vector now we
are loop directly on bit vector.
The second one is that we don't unconditionally clear content of internal list, but
overwrite old elements and clear leftovers afterwards.